requirements.txt
# Dev environment
pip
autopep8

# App
flask
autopep8==2.3.2
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector==2.2.9
packaging==24.2
pycodestyle==2.12.1
PyMySQL==1.1.1
Werkzeug==3.1.3
google-cloud-vision
